Add files via upload
This project aims to develop a machine learning model capable of detecting whether a person is wearing a face mask or not in an image. The model is built using convolutional neural networks (CNNs), a type of deep learning algorithm particularly effective for image classification tasks.

Key Components:

Dataset: The project utilizes a dataset consisting of images with individuals wearing face masks and images without face masks. Each image is labeled with the corresponding class (with mask or without mask).

Model Architecture: The neural network model architecture comprises convolutional layers for feature extraction, followed by pooling layers for dimensionality reduction, and dense layers for classification. Dropout layers are added for regularization to prevent overfitting.

Training: The model is trained using the labeled dataset, where it learns to distinguish between images with and without face masks. The training process involves optimizing the model parameters using the Adam optimizer and minimizing the cross-entropy loss function.

Evaluation: The trained model is evaluated on a separate validation dataset to assess its performance in terms of accuracy and other relevant metrics.

Deployment: Once trained and evaluated, the model can be deployed to detect face masks in new images or real-time video streams. It can be integrated into applications or systems for various purposes, such as monitoring compliance with face mask regulations in public spaces.